[SIG-2348] Refactored onSubmit, onCancel UserForm logic
